# openapiv2/notes/v1/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Notes",
    "version": "1.0",
    "contact": {
      "name": "bbengfort",
      "url": "https://github.com/bbengfort/notes",
      "email": "info@bengfort.com"
    },
    "license": {
      "name": "BSD 3-Clause License",
      "url": "https://github.com/bbengfort/notes/LICENSE"
    }
  },
  "tags": [
    {
      "name": "NoteService"
    }
  ],
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/notes": {
      "get": {
        "operationId": "NoteService_Fetch",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Notebook"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string",
              "format": "uint64"
            },
            "collectionFormat": "multi"
          },
          {
            "name": "author",
            "in": "query",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "after",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "private",
            "in": "query",
            "required": false,
            "type": "boolean"
          }
        ],
        "tags": [
          "NoteService"
        ]
      },
      "post": {
        "operationId": "NoteService_Create",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Notebook"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1Note"
            }
          }
        ],
        "tags": [
          "NoteService"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Error": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int64"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "v1Note": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uint64"
        },
        "timestamp": {
          "type": "string"
        },
        "author": {
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "private": {
          "type": "boolean"
        }
      }
    },
    "v1Notebook": {
      "type": "object",
      "properties": {
        "error": {
          "$ref": "#/definitions/v1Error"
        },
        "notes": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/v1Note"
          }
        }
      }
    }
  }
}